fix: add git to container to ensure reproducibility
Removing `package-lock.json` would cause subsequent full builds of
the container to fail. This is because git is a requirement of
some dependencies of packages listed in `package.json`. Git and
curl have now been added to the Dockerfile. In addition, the
`package-lock.json` has been regenerated with the latest dependency
versions.